Refactor fetching logic
The new keg.core.fetcher module brings a Fetcher class which manages
fetching and verifying items for a specific version

Running `ngdp fetch <version>` will now first fetch all metadata for
all versions first, then fetch all data for all versions. Each version
is iterated one at a time.

The FetchManager is instanciated with a version and contains a bunch of
queues (for config, indices, data, etc).
FetchManager exposes two methods: fetch_metadata() and fetch_data().
They should be called in that order. The methods fill the queues and
yield a "Drain" of the queue as required by the fetch process.

The Drain class has a drain() method which itself is a generator of
FetchDirective objects. Those directives have a fetch() method which can
be called to execute the directive (= attempt fetching over the wire).

The high level API looks like this:

    fetcher = FetchManager(version)
    for queue in fetcher.fetch_metadata():
        for item in queue.drain():
            item.fetch()

This allows finer-grained control over the fetch flow.

This rework was driven by multiple needs:

1. Fetch logic needed to be decoupled from the program and moved to the
   library for programmatic access.
2. The upcoming decryption support needed a more centralized place that
   is aware of all the fetch state.
3. Better progress bars with download speed support require supporting
   chunking of arbitrary requests and iterating over those chunks. The
   likely approach will be to add a fetch_chunks() method on the
   FetchDirective class which returns a generator of those chunks.

Closes #6